service: db: add direct reference of storage_service to migration_manager
There is already indirect link via migration_notifier. This is needed
because in following commits we'll split update_tablet_metadata into
steps so that it can be used in non-yielding atomic schema update code.

Moreover it's misleading to use decoupled notification but require
strict order of the call (after tables are dropped but before created)

Original on_update_tablet_metadata() notification was not removed,
potentially it could be usefull in the future, maybe driver will need
to be informed about this.